pyproject.toml
[tool.poetry]
name = "dear"
version = "0.1.0"
description = ""
authors = ["Cecil <cecil.ash@karbonhq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.8.13"
streamlit = "^1.22.0"
seaborn = "^0.12.2"
plotly = "^5.14.1"
pandas = "^2.0.1"
numpy = "^1.24.3"
snowflake-connector-python = "^3.0.3"
snowflake-snowpark-python = "^1.4.0"
snowflake = "^0.0.3"
uuid = "^1.30"
python-dotenv = "^1.0.0"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
